backend/requirements.txt
pillow
numpy
matplotlib
scikit-learn
pandas
fastapi
uvicorn
python-multipart